issue=1119: more code style fix for coverity

+++ src/leveldb/db/log_writer.cc
@@ -51,7 +51,7 @@ Status Writer::AddRecord(const Slice& slice) {
     }
 
     // Invariant: we never leave < kHeaderSize bytes in a block.
-    assert(kBlockSize - block_offset_ - kHeaderSize >= 0);
+    assert(kBlockSize - block_offset_ - static_cast<const int>(kHeaderSize) >= 0);
 
     const size_t avail = kBlockSize - block_offset_ - kHeaderSize;
     const size_t fragment_length = (left < avail) ? left : avail;
